requirements.txt
Flask>=2.0,<4.0
Flask-WTF>=1.0,<2.0
Flask-Login>=0.6,<0.8
Flask-SQLAlchemy>=3.0,<4.0
Werkzeug>=2.0,<4.0
email-validator>=1.0,<3.0
Pillow>=8.0,<12.0
torch>=2.0
torchvision>=0.15.0
torchaudio>=2.0